requirements.txt
playwright==1.44.0
gspread==6.1.2
google-auth==2.29.0
openai==1.35.0
python-dotenv==1.0.1
beautifulsoup4==4.12.3
httpx==0.27.0
